Models fixes

+++ LazyStockDiaryApi/Models/HistoricalEod.cs
@@ -19,7 +19,7 @@ namespace LazyStockDiaryApi.Models
         public double? Close { get; set; }
         public double? High { get; set; }
         public double? Low { get; set; }
-        public int? Volume { get; set; }
+        public ulong? Volume { get; set; }
     }
 }
